Move debounce effect setup to dedicated supportDebounce.js
